wip: Dedupe business logic and move to separate files in the `overrides` folder
